Yong Deng-Homework 2
This program is a simple search engine which can perform AND and OR
searches. The user can enter multiple words and boolean operators (and,
or). If user only entered the "or" operator, an OR search will be
performed. Otherwise, an AND search will be performed. An empty query
or a one that only contains boolean operators is invalid. The user will
be prompted to enter query again.

Sample outputs have been pasted at the end of the source file.